Add pin_gecko_rev_yml script
pin_gecko_rev_yml is useful for submitting try-c-c jobs that pin to
something other than mozilla-central:tip.

It accepts one optional argument, the name of a mozilla repo. Then, it
gets the revision of HEAD from the mozilla checkout (at .. from cwd)
and updates .gecko_rev.yml with the revision and mozilla repo URL.

For example, if you are working on a patch for something in autoland,
in $topsrcdir:
- hg pull autoland
- hg up -r <autoland rev>
Then in $topsrcdir/comm:
- pin_gecko_rev_yml autoland

.gecko_rev.yml will be updated with autoland's URL in
GECKO_HEAD_REPOSITORY and <autoland rev> for GECKO_HEAD_REV.

Push to try-c-c as you normally do. If using the push-to-try
Mercurial extension, you don't need to commit .gecko_rev.yml.